add: hooks state
