+++ src/components/page/StreamingPage/page.tsx
@@ -1,7 +1,7 @@
 "use client";
-import React, { useCallback, useRef, useState } from "react";
+import React, { useCallback, useState } from "react";
+
 import { useSockets } from "@/app/providers/socket";
-import { guardUndef } from "@/utils/guardUndef";
 
 export const StreamingPage = () => {
     const { socket } = useSockets();
